requirements.txt
flask
python-docx
PyPDF2
huggingface_hub
transformers
